+++ resolwe_bio/processes/samtools/samtools_bedcov.py
@@ -64,7 +64,7 @@ class SamtoolsBedcov(Process):
     requirements = {
         "expression-engine": "jinja",
         "executor": {
-            "docker": {"image": "public.ecr.aws/s4q6j6e8/resolwebio/common:3.1.0"},
+            "docker": {"image": "public.ecr.aws/genialis/resolwebio/common:4.1.1"},
         },
         "resources": {
             "cores": 2,
@@ -73,7 +73,7 @@ class SamtoolsBedcov(Process):
     }
     category = "Samtools"
     data_name = "{{ bam|name|default('?') }}"
-    version = "1.0.0"
+    version = "1.0.1"
     entity = {"type": "sample"}
     scheduling_class = SchedulingClass.BATCH